+++ tests/unit-ops.zig
@@ -33,6 +33,8 @@ fn assertEq(val1: f32, val2: f32) !void {
     }
 }
 
+// TODO: Refactor these Tensor things like in the other tests
+
 // TODO: --loop option
 pub fn main() !void {
     var gpa = std.heap.GeneralPurposeAllocator(.{}){};
@@ -55,10 +57,10 @@ pub fn main() !void {
         }
     }
 
-    const a_size: u32 = 6;
-    const z_size: u32 = 5;
-    const y_size: u32 = 4;
-    const x_size: u32 = 3;
+    const a_size: usize = 6;
+    const z_size: usize = 5;
+    const y_size: usize = 4;
+    const x_size: usize = 3;
     var tensor1 = try Tensor.alloc(allocator, a_size, z_size, y_size, x_size, null);
     defer tensor1.free(allocator);
     var tensor2 = try Tensor.alloc(allocator, a_size, z_size, y_size, x_size, null);
